runtime: Continue kata-runtime darwin work
Continued work to get kata-runtime to build. Now there's a single
thing holding it back which is that libcontainer's mount definition
uses a Linux only struct inside which will need to be fixed in runc.
