+++ src/runtime/cmd/kata-runtime/kata-check_linux_arm64.go
@@ -96,7 +96,6 @@ func archHostCanCreateVMContainer(hypervisorType vc.HypervisorType) error {
 // hostIsVMContainerCapable checks to see if the host is theoretically capable
 // of creating a VM container.
 func hostIsVMContainerCapable(details vmContainerCapableDetails) error {
-
 	_, err := getCPUInfo(details.cpuInfoFile)
 	if err != nil {
 		return err
