+++ src/runtime/virtcontainers/factory/template/template_linux.go
@@ -16,7 +16,6 @@ import (
 	pb "github.com/kata-containers/kata-containers/src/runtime/protocols/cache"
 	vc "github.com/kata-containers/kata-containers/src/runtime/virtcontainers"
 	"github.com/kata-containers/kata-containers/src/runtime/virtcontainers/factory/base"
-	"github.com/sirupsen/logrus"
 )
 
 type template struct {
